refactoring : common 패키지 안으로 util 패키지 이동
여러 도메인에서 관리할 Enum 클래스가 존재하기에 Common 패키지 생성
