+++ src/main/java/com/mtvs/devlinkbackend/common/util/EpicGamesJWKCache.java
@@ -1,4 +1,4 @@
-package com.mtvs.devlinkbackend.util;
+package com.mtvs.devlinkbackend.common.util;
 
 import com.nimbusds.jose.jwk.JWKSet;
 import org.springframework.scheduling.annotation.Scheduled;
