+++ src/main/java/com/mtvs/devlinkbackend/member/command/service/MemberService.java
@@ -1,6 +1,6 @@
 package com.mtvs.devlinkbackend.member.command.service;
 
-import com.mtvs.devlinkbackend.member.command.model.entity.AcceptStatus;
+import com.mtvs.devlinkbackend.common.model.AcceptStatus;
 import com.mtvs.devlinkbackend.member.command.model.entity.Member;
 import com.mtvs.devlinkbackend.member.command.repository.MemberRepository;
 import com.mtvs.devlinkbackend.member.query.view.response.MemberStatusResponseDTO;
