+++ src/main/java/com/mtvs/devlinkbackend/user/command/controller/EpicBusinessCommandController.java
@@ -3,7 +3,7 @@ package com.mtvs.devlinkbackend.user.command.controller;
 import com.mtvs.devlinkbackend.user.command.model.dto.request.BusinessRequestDTO;
 import com.mtvs.devlinkbackend.user.query.model.dto.response.BusinessSingleResponseDTO;
 import com.mtvs.devlinkbackend.user.command.service.EpicBusinessService;
-import com.mtvs.devlinkbackend.util.JwtUtil;
+import com.mtvs.devlinkbackend.common.util.JwtUtil;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
