+++ src/main/java/com/mtvs/devlinkbackend/user/command/model/entity/Dev.java
@@ -1,7 +1,7 @@
 package com.mtvs.devlinkbackend.user.command.model.entity;
 
 import com.fasterxml.jackson.annotation.JsonManagedReference;
-import com.mtvs.devlinkbackend.util.converter.StringListConverter;
+import com.mtvs.devlinkbackend.common.util.converter.StringListConverter;
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
