+++ src/main/java/com/mtvs/devlinkbackend/user/query/controller/EpicDevQueryController.java
@@ -3,7 +3,7 @@ package com.mtvs.devlinkbackend.user.query.controller;
 import com.mtvs.devlinkbackend.user.query.model.dto.response.DevPagingResponseDTO;
 import com.mtvs.devlinkbackend.user.query.model.dto.response.DevSingleResponseDTO;
 import com.mtvs.devlinkbackend.user.query.service.EpicDevViewService;
-import com.mtvs.devlinkbackend.util.JwtUtil;
+import com.mtvs.devlinkbackend.common.util.JwtUtil;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
