+++ src/test/java/com/mtvs/devlinkbackend/crud/MemberCRUDTest.java
@@ -1,6 +1,6 @@
 package com.mtvs.devlinkbackend.crud;
 
-import com.mtvs.devlinkbackend.member.command.model.entity.AcceptStatus;
+import com.mtvs.devlinkbackend.common.model.AcceptStatus;
 import com.mtvs.devlinkbackend.member.command.model.entity.Member;
 import com.mtvs.devlinkbackend.member.command.repository.MemberRepository;
 import com.mtvs.devlinkbackend.member.command.service.MemberService;
